envoy-control-plane/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
httpx==0.27.0
pyyaml==6.0.1
opentelemetry-sdk>=1.27.0
opentelemetry-instrumentation-fastapi>=0.48b0
opentelemetry-exporter-otlp-proto-http>=1.27.0
setuptools>=69.0.0
